feat(AVO-1723): upgrade react-router v6 to support server side rendering
https://meemoo.atlassian.net/browse/AVO-1723

+++ src/collection/components/CollectionOrBundleEdit.tsx
@@ -26,13 +26,12 @@ import React, {
 	useState,
 } from 'react';
 import { Helmet } from 'react-helmet';
-import { matchPath, Redirect, withRouter } from 'react-router';
+import { matchPath, Navigate, useMatch, useNavigate } from 'react-router';
 import { compose } from 'redux';
 
 import { ItemsService } from '../../admin/items/items.service';
 import { reorderBlockPositions, setBlockPositionToIndex } from '../../assignment/assignment.helper';
 import { AssignmentService } from '../../assignment/assignment.service';
-import { type DefaultSecureRouteProps } from '../../authentication/components/SecuredRoute';
 import { PermissionService } from '../../authentication/helpers/permission-service';
 import { redirectToClientPage } from '../../authentication/helpers/redirects/redirect-to-client-page';
 import { APP_PATH, GENERATE_SITE_TITLE } from '../../constants';
@@ -120,15 +119,18 @@ import { DeleteMyselfFromCollectionContributorsConfirmModal } from './modals/Del
 
 import './CollectionOrBundleEdit.scss';
 
-const CollectionOrBundleEdit: FC<
-	CollectionOrBundleEditProps &
-		DefaultSecureRouteProps<{ id: string; tabId: CollectionCreateUpdateTab | undefined }> &
-		UserProps
-> = ({ type, history, match, commonUser }) => {
+const CollectionOrBundleEdit: FC<CollectionOrBundleEditProps & UserProps> = ({
+	type,
+	commonUser,
+}) => {
 	const { tText, tHtml } = useTranslation();
+	const navigateFunc = useNavigate();
+	const match = useMatch<'id' | 'tabId', string>(APP_PATH.COLLECTION_EDIT.route);
+
+	const collectionId = match?.params.id;
+	const tabId = match?.params.tabId as CollectionCreateUpdateTab | undefined;
 
 	// State
-	const [collectionId] = useState<string>(match.params.id);
 	const [currentTab, setCurrentTab] = useState<CollectionCreateUpdateTab | null>(null);
 	const [isOptionsMenuOpen, setIsOptionsMenuOpen] = useState<boolean>(false);
 	const [isSavingCollection, setIsSavingCollection] = useState<boolean>(false);
@@ -165,8 +167,8 @@ const CollectionOrBundleEdit: FC<
 
 	const isCollection = type === 'collection';
 	const isAdmin = isCollection
-		? commonUser.permissions?.includes(PermissionName.EDIT_ANY_COLLECTIONS)
-		: commonUser.permissions?.includes(PermissionName.EDIT_ANY_BUNDLES);
+		? commonUser?.permissions?.includes(PermissionName.EDIT_ANY_COLLECTIONS)
+		: commonUser?.permissions?.includes(PermissionName.EDIT_ANY_BUNDLES);
 	const noRightsError = useMemo(
 		() =>
 			({
@@ -185,6 +187,9 @@ const CollectionOrBundleEdit: FC<
 
 	const releaseCollectionEditStatus = useCallback(async () => {
 		try {
+			if (!collectionId) {
+				return;
+			}
 			await CollectionService.releaseCollectionEditStatus(collectionId);
 		} catch (err) {
 			if ((err as CustomError)?.innerException?.additionalInfo?.statusCode !== 409) {
@@ -199,11 +204,14 @@ const CollectionOrBundleEdit: FC<
 
 	const updateCollectionEditor = useCallback(async () => {
 		try {
+			if (!collectionId) {
+				return;
+			}
 			await CollectionService.updateCollectionEditor(collectionId);
 		} catch (err) {
 			redirectToClientPage(
 				buildLink(APP_PATH.COLLECTION_DETAIL.route, { id: collectionId }),
-				history
+				navigateFunc
 			);
 
 			if ((err as CustomError).innerException?.additionalInfo?.statusCode === 409) {
@@ -223,7 +231,7 @@ const CollectionOrBundleEdit: FC<
 				);
 			}
 		}
-	}, [collectionId, history, releaseCollectionEditStatus, tHtml]);
+	}, [collectionId, navigateFunc, releaseCollectionEditStatus, tHtml]);
 
 	const updateCollectionEditorWithLoading = useCallback(async () => {
 		if (!isCollection) {
@@ -463,10 +471,10 @@ const CollectionOrBundleEdit: FC<
 	const shouldDeleteSelfFromCollection = isContributor && !permissions.canDelete;
 
 	const { data: bundlesContainingCollection } = useGetCollectionsOrBundlesContainingFragment(
-		collectionId,
+		collectionId as string,
 		BundleSortProp.title,
 		OrderDirection.asc,
-		{ enabled: !!collectionState.currentCollection }
+		{ enabled: !!collectionId && !!collectionState.currentCollection }
 	);
 
 	useEffect(() => {
@@ -494,6 +502,9 @@ const CollectionOrBundleEdit: FC<
 
 	const checkPermissionsAndGetCollection = useCallback(async (): Promise<void> => {
 		try {
+			if (!collectionId) {
+				return;
+			}
 			const permissionObj = await PermissionService.checkPermissions(
 				{
 					canEdit: [
@@ -680,14 +691,14 @@ const CollectionOrBundleEdit: FC<
 
 	// react to route changes by navigating back wih the browser history back button
 	useEffect(() => {
-		setCurrentTab(match.params.tabId || CollectionCreateUpdateTab.CONTENT);
-	}, [match.params.tabId]);
+		setCurrentTab(tabId || CollectionCreateUpdateTab.CONTENT);
+	}, [tabId]);
 
 	// Change page on tab selection
 	const selectTab = (selectedTab: ReactText) => {
 		const tabName = String(selectedTab) as CollectionCreateUpdateTab;
 		navigate(
-			history,
+			navigateFunc,
 			isCollection ? APP_PATH.COLLECTION_EDIT_TAB.route : APP_PATH.BUNDLE_EDIT_TAB.route,
 			{ id: collectionId, tabId: tabName }
 		);
@@ -784,7 +795,8 @@ const CollectionOrBundleEdit: FC<
 				!!collectionState.currentCollection?.management_quality_check?.[0]) &&
 			!collectionState.currentCollection?.management_language_check?.[0]?.assignee_profile_id
 		) {
-			history.push(
+			navigate(
+				navigateFunc,
 				buildLink(
 					isCollection
 						? APP_PATH.COLLECTION_EDIT_TAB.route
@@ -807,6 +819,9 @@ const CollectionOrBundleEdit: FC<
 	};
 
 	const updateCollection = async (checkValidation = true) => {
+		if (!collectionId) {
+			return;
+		}
 		if (isNil(collectionState.currentCollection)) {
 			console.error('Current collection state is nil');
 			ToastService.danger(
@@ -921,18 +936,21 @@ const CollectionOrBundleEdit: FC<
 	};
 
 	const handleDeleteCollection = async () => {
+		if (!collectionId) {
+			return;
+		}
 		await deleteCollection(
 			collectionId,
 			commonUser,
 			isCollection,
 			async () => await CollectionService.deleteCollectionOrBundle(collectionId),
-			() => navigate(history, APP_PATH.WORKSPACE_TAB.route, { tabId: COLLECTIONS_ID })
+			() => navigate(navigateFunc, APP_PATH.WORKSPACE_TAB.route, { tabId: COLLECTIONS_ID })
 		);
 	};
 
 	const handleDeleteSelfFromCollection = async () => {
 		await deleteSelfFromCollection(collectionId, commonUser, () =>
-			navigate(history, APP_PATH.WORKSPACE_TAB.route, { tabId: COLLECTIONS_ID })
+			navigate(navigateFunc, APP_PATH.WORKSPACE_TAB.route, { tabId: COLLECTIONS_ID })
 		);
 	};
 
@@ -976,10 +994,10 @@ const CollectionOrBundleEdit: FC<
 							? APP_PATH.COLLECTION_DETAIL.route
 							: APP_PATH.BUNDLE_DETAIL.route,
 						{
-							id: match.params.id,
+							id: collectionId,
 						}
 					),
-					history
+					navigateFunc
 				);
 				break;
 
@@ -1228,7 +1246,7 @@ const CollectionOrBundleEdit: FC<
 			buildLink(APP_PATH.COLLECTION_DETAIL.route, {
 				id: collectionId,
 			}),
-			history
+			navigateFunc
 		);
 	};
 
@@ -1258,7 +1276,6 @@ const CollectionOrBundleEdit: FC<
 						<CollectionOrBundleEditAdmin
 							collection={collectionState.currentCollection}
 							changeCollectionState={changeCollectionState}
-							history={history}
 							onFocus={() => setUnsavedChanges(true)}
 						/>
 					);
@@ -1267,7 +1284,6 @@ const CollectionOrBundleEdit: FC<
 						<CollectionOrBundleEditActualisation
 							collection={collectionState.currentCollection}
 							changeCollectionState={changeCollectionState}
-							history={history}
 							onFocus={() => setUnsavedChanges(true)}
 						/>
 					);
@@ -1276,7 +1292,6 @@ const CollectionOrBundleEdit: FC<
 						<CollectionOrBundleEditQualityCheck
 							collection={collectionState.currentCollection}
 							changeCollectionState={changeCollectionState}
-							history={history}
 							onFocus={() => setUnsavedChanges(true)}
 						/>
 					);
@@ -1285,7 +1300,6 @@ const CollectionOrBundleEdit: FC<
 						<CollectionOrBundleEditMarcom
 							collection={collectionState.currentCollection}
 							changeCollectionState={changeCollectionState}
-							history={history}
 							onFocus={() => setUnsavedChanges(true)}
 						/>
 					);
@@ -1306,6 +1320,9 @@ const CollectionOrBundleEdit: FC<
 		PermissionName.ADD_ASSIGNMENT_TO_BUNDLE
 	);
 	const renderHeaderButtons = () => {
+		if (!collectionId) {
+			return null;
+		}
 		const COLLECTION_DROPDOWN_ITEMS = [
 			...createDropdownMenuItem(
 				collectionId,
@@ -1496,6 +1513,9 @@ const CollectionOrBundleEdit: FC<
 	};
 
 	const renderHeaderButtonsMobile = () => {
+		if (!collectionId) {
+			return;
+		}
 		const COLLECTION_DROPDOWN_ITEMS = [
 			...createDropdownMenuItem(
 				collectionId,
@@ -1665,6 +1685,7 @@ const CollectionOrBundleEdit: FC<
 				)}
 
 				{collectionState.currentCollection &&
+					!!collectionId &&
 					renderMobileDesktop({
 						mobile: (
 							<ShareModal
@@ -1772,7 +1793,7 @@ const CollectionOrBundleEdit: FC<
 						warningMessage={tHtml(
 							'collection/components/collection-or-bundle-edit___door-inactiviteit-zal-de-collectie-zichzelf-sluiten'
 						)}
-						currentPath={history.location.pathname}
+						currentPath={location.pathname}
 						editPath={APP_PATH.COLLECTION_EDIT_TAB.route}
 						onForcedExit={onForcedExitPage}
 					/>
@@ -1849,9 +1870,9 @@ const CollectionOrBundleEdit: FC<
 		);
 	};
 
-	if (matchPath(location.pathname, { path: APP_PATH.BUNDLE_EDIT.route, exact: true })) {
+	if (matchPath(location.pathname, APP_PATH.BUNDLE_EDIT.route)) {
 		return (
-			<Redirect
+			<Navigate
 				to={buildLink(APP_PATH.BUNDLE_EDIT_TAB.route, {
 					id: collectionId,
 					tabId: CollectionCreateUpdateTab.CONTENT,
@@ -1860,9 +1881,9 @@ const CollectionOrBundleEdit: FC<
 		);
 	}
 
-	if (matchPath(location.pathname, { path: APP_PATH.COLLECTION_EDIT.route, exact: true })) {
+	if (matchPath(location.pathname, APP_PATH.COLLECTION_EDIT.route)) {
 		return (
-			<Redirect
+			<Navigate
 				to={buildLink(APP_PATH.COLLECTION_EDIT_TAB.route, {
 					id: collectionId,
 					tabId: CollectionCreateUpdateTab.CONTENT,
@@ -1898,7 +1919,4 @@ const CollectionOrBundleEdit: FC<
 	);
 };
 
-export default compose(
-	withRouter,
-	withUser
-)(CollectionOrBundleEdit) as FC<CollectionOrBundleEditProps>;
+export default compose(withUser)(CollectionOrBundleEdit) as FC<CollectionOrBundleEditProps>;
